[Test][Benchmark] fold BenchmarkReport.record tests into test_benchmark_base
One module, one test file; the autouse fixture becomes explicit
usefixtures so it cannot leak onto unrelated tests.
